requirements.txt
asgiref==3.4.1
Django==3.0
django-js-asset==1.2.2
django-mptt==0.13.4
django-phonenumber-field==6.0.0
django-widget-tweaks==1.4.9
djangorestframework==3.13.1
phonenumbers==8.12.39
Pillow==8.4.0
pytz==2021.3
sqlparse==0.4.2
